requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20210324.2=h2531618_0
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py39h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
alsa-lib=1.2.3.2=h166bdaf_0
astor=0.8.1=py39h06a4308_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=hf32fb01_1
certifi=2022.6.15=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
cryptography=3.4.8=py39hd23ed53_0
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
ffmpeg=4.3.2=hca11adc_0
flatbuffers=2.0.0=h2531618_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py39h7f8727e_0
gast=0.4.0=pyhd3eb1b0_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
glib=2.68.4=h9c3ff4c_0
glib-tools=2.68.4=h9c3ff4c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyhd3eb1b0_0
graphite2=1.3.14=h295c915_1
grpcio=1.42.0=py39hce63b2e_0
gst-plugins-base=1.18.5=hf529b03_0
gstreamer=1.18.5=h76c114f_0
h5py=3.6.0=py39ha0f2276_0
harfbuzz=2.9.1=h83ec7ef_1
hdf5=1.10.6=hb1b8bf9_0
icu=68.1=h2531618_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py39h06a4308_0
ipython=8.3.0=py39h06a4308_0
jasper=1.900.1=hd497a04_4
jedi=0.18.1=py39h06a4308_1
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h7f8727e_0
jupyter_client=7.2.2=py39h06a4308_0
jupyter_core=4.10.0=py39h06a4308_0
keras=2.6.0=pyhd3eb1b0_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=py39h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=12_linux64_mkl
libclang=11.1.0=default_ha53f305_1
libcurl=7.82.0=h0b77cf5_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.10=h9b69904_4
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.68.4=h3e27bee_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm11=11.1.0=h3826bc1_1
libnghttp2=1.46.0=hce63b2e_0
libogg=1.3.5=h27cfd23_1
libopencv=4.5.1=py39hf3c14e0_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=13.3=hd57d9b9_0
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libvorbis=1.3.7=h7b6447c_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
llvmlite=0.38.0=py39h4ff587b_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py39h06a4308_0
matplotlib=3.5.2=py39hf3d152e_0
matplotlib-base=3.5.2=py39h700656a_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=5.2.0=py39h7f8727e_2
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.25=ha770c72_2
mysql-libs=8.0.25=hfa10184_2
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numba=0.55.1=py39h51133e4_0
numexpr=2.8.3=py39h807cd23_0
numpy=1.21.5=py39h6c91a56_3
numpy-base=1.21.5=py39ha15fc14_3
oauthlib=3.2.0=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
opencv=4.5.1=py39hf3d152e_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39h295c915_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=7.2.0=py39h6f3857e_2
pip=21.2.4=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.14.0=py39h2531618_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-opencv=4.5.1=py39hef51801_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.1.0=py39h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.12.3=py39h03dd644_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2022.1=py39h06a4308_0
pyzmq=22.3.0=py39h295c915_2
qt=5.12.9=hda022c4_4
readline=8.1.2=h7f8727e_1
requests=2.28.0=py39h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.1.1=py39h4037b75_0
scipy=1.7.3=py39hc147768_0
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.38.5=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.5.0=hd09550d_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py39hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.6.0=mkl_py39haac40d1_0
tensorflow-base=2.6.0=mkl_py39h3d85931_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
termcolor=1.1.0=py39h06a4308_1
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
tornado=6.1=py39h27cfd23_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.13.3=py39h7f8727e_2
x264=1!161.3030=h7f98852_1
xz=5.2.5=h7f8727e_1
yarl=1.6.3=py39h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py39h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0

